NB-A2-07: Review fixes
