+++ src/data/Transactions/Transactions.types.ts
@@ -1,13 +1,3 @@
-import { JSX } from 'react';
-
-/**
- * A union type representing all possible transaction statuses.
- * - `'Completed'`: The transaction was successful.
- * - `'Cancelled'`: The transaction was cancelled.
- * - `'In progress'`: The transaction is currently being processed.
- */
-export type TransactionStatus = 'Completed' | 'Cancelled' | 'In progress';
-
 /**
  * Represents a single financial transaction.
  */
@@ -30,41 +20,5 @@ export type Transaction = {
     amount: number;
 
     /** Current status of the transaction */
-    status: TransactionStatus;
-};
-
-/**
- * Defines the structure of a column used in a data table.
- *
- * @template T - The type of the data rows.
- */
-export type TableColumn<T> = {
-    /** The key of the field in the data object to display */
-    key: keyof T;
-
-    /** Column header label */
-    label: string;
-
-    /** Text alignment for the cell.*/
-    align?: 'left' | 'right' | 'center';
-
-    /**
-     * Optional custom rendering function for the cell content.
-     */
-    render: (value: T[keyof T], row: T) => JSX.Element;
-};
-
-/**
- * Defines the props for the `TransactionsTable` component,
- * which is responsible for rendering a table of transaction data.
- */
-export type TransactionsTableProps = {
-    /** Optional title to be displayed above the table */
-    title: string;
-
-    /** Columns configuration for the table */
-    columns: TableColumn<Transaction>[];
-
-    /** The data rows (transactions) to display in the table */
-    rows: Transaction[];
+    status: 'Completed' | 'Cancelled' | 'In progress';
 };
